Refactor training script
