+++ utils/__init__.py
@@ -1,3 +1,4 @@
 from utils.loading_data import *
 from utils.preprocess import *
 from utils.data_loader import *
+from utils.help_training import *
